+++ src/main/java/org/mesdag/advjs/mixin/BlockBehavior$BlockStateBaseMixin.java
@@ -15,15 +15,15 @@ import org.spongepowered.asm.mixin.injection.At;
 import org.spongepowered.asm.mixin.injection.Inject;
 import org.spongepowered.asm.mixin.injection.callback.CallbackInfoReturnable;
 
-import static org.mesdag.advjs.util.Data.LOCK_INTERACT;
+import static org.mesdag.advjs.util.Data.LOCK_BLOCK_INTERACT;
 
 @Mixin(BlockBehaviour.BlockStateBase.class)
 public abstract class BlockBehavior$BlockStateBaseMixin {
     @Shadow public abstract Block getBlock();
 
     @Inject(method = "use", at = @At("HEAD"), cancellable = true)
-    private void advJS$interrupt(Level level, Player player, InteractionHand hand, BlockHitResult result, CallbackInfoReturnable<InteractionResult> cir) {
-        if (level.isClientSide || !LOCK_INTERACT.containsKey(getBlock())) return;
-        if (AdvHelper.advNotDone((ServerPlayer) player, LOCK_INTERACT.get(getBlock()))) cir.setReturnValue(InteractionResult.PASS);
+    private void advJS$pass(Level level, Player player, InteractionHand hand, BlockHitResult result, CallbackInfoReturnable<InteractionResult> cir) {
+        if (level.isClientSide) return;
+        if (!AdvHelper.advDone((ServerPlayer) player, LOCK_BLOCK_INTERACT.get(getBlock()))) cir.setReturnValue(InteractionResult.PASS);
     }
 }